requirements.txt
Django==1.11.6
django-filter==2.1.0
djangorestframework==3.9.0
olefile==0.44
Pillow==4.2.0
pytz==2017.2
